Remove 'Global Mod Settings' button
This button is used to access GuiAPI to configure the mods, but it crashes
when used with SK's Launcher, see:

http://www.minecraftforum.net/topic/1104030-launcher-sks-launcher-v315-easily-install-mods-update-mc-open-source/page__st__40#entry13649982

On SMP, configuring the mods is not needed for players (at least not with Mo' Creatures),
and the crash confuses users. This change to GuiAPI allows Mo' Creatures to function,
without offering any configurability (which is not needed in all cases).

https://github.com/downloads/mushroomhostage/GuiAPI/GuiAPI-0.14.2-1.2.5+no-options.zip